// types/Enums.ts
export enum DAY {
  All = 'All',
  Sun = 'Sun',
  Mon = 'Mon',
  Tue = 'Tue',
  Wed = 'Wed',
  Thu = 'Thu',
  Fri = 'Fri',
  Sat = 'Sat',
}

export enum STATE {
  All = 'All',
  WA = 'WA',
  NT = 'NT',
  SA = 'SA',
  QLD = 'QLD',
  NSW = 'NSW',
  ACT = 'ACT',
  VIC = 'VIC',
  TAS = 'TAS',
}

export enum PRODUCT {
  All = 'All',
}